Adding stubs to permit compilation on other platforms
