+++ internal/web/apiv1_os_specific_freebsd.go
@@ -1,0 +1,13 @@
+package web
+
+import (
+	"errors"
+	"fmt"
+	"unsafe"
+)
+
+func getResult(counter string) CounterResult {
+	returnvalue := CounterResult{}
+
+	return returnvalue
+}
